requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-celery==3.3.1
django-cors-headers==3.5.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
djoser==2.0.5
idna==2.10
kombu==3.0.37
Pillow==8.0.0
pytz==2020.1
requests==2.24.0
six==1.15.0
spotipy==2.16.0
sqlparse==0.4.1
urllib3==1.25.11
